Move _eeid_patch_memory()
