+++ enclave/core/sgx/calls.c
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 #include "../calls.h"
-#include <openenclave/bits/eeid.h>
 #include <openenclave/bits/sgx/sgxtypes.h>
 #include <openenclave/corelibc/stdlib.h>
 #include <openenclave/corelibc/string.h>
@@ -11,7 +10,6 @@
 #include <openenclave/internal/allocator.h>
 #include <openenclave/internal/atomic.h>
 #include <openenclave/internal/calls.h>
-#include <openenclave/internal/eeid.h>
 #include <openenclave/internal/fault.h>
 #include <openenclave/internal/globals.h>
 #include <openenclave/internal/jump.h>
@@ -135,40 +133,6 @@ extern bool oe_disable_debug_malloc_check;
 **==============================================================================
 */
 
-#ifdef OE_WITH_EXPERIMENTAL_EEID
-extern volatile const oe_sgx_enclave_properties_t oe_enclave_properties_sgx;
-extern oe_eeid_t* oe_eeid;
-
-int _is_eeid_base_image(const volatile oe_sgx_enclave_properties_t* properties)
-{
-    return properties->header.size_settings.num_heap_pages == 0 &&
-           properties->header.size_settings.num_stack_pages == 0 &&
-           properties->header.size_settings.num_tcs == 1;
-}
-
-static oe_result_t _eeid_patch_memory()
-{
-    oe_result_t r = OE_OK;
-
-    if (_is_eeid_base_image(&oe_enclave_properties_sgx))
-    {
-        uint8_t* enclave_base = (uint8_t*)__oe_get_enclave_base();
-        uint8_t* heap_base = (uint8_t*)__oe_get_heap_base();
-        oe_eeid_marker_t* marker = (oe_eeid_marker_t*)heap_base;
-        oe_eeid = (oe_eeid_t*)(enclave_base + marker->offset);
-
-        // Wipe the marker page
-        memset(heap_base, 0, OE_PAGE_SIZE);
-
-        // Re-initialize the memory allocator with updated heap size
-        oe_allocator_init(
-            (void*)__oe_get_heap_base(), (void*)__oe_get_heap_end());
-    }
-
-    return r;
-}
-#endif
-
 /*
 **==============================================================================
 **
@@ -196,10 +160,6 @@ static oe_result_t _handle_init_enclave(uint64_t arg_in)
         if (_once == false)
         {
             oe_enclave_t* enclave = (oe_enclave_t*)arg_in;
-
-#ifdef OE_WITH_EXPERIMENTAL_EEID
-            OE_CHECK(_eeid_patch_memory());
-#endif
 
 #ifdef OE_USE_BUILTIN_EDL
             /* Install the common TEE ECALL function table. */
