+++ enclave/core/sgx/init.c
@@ -2,12 +2,17 @@
 // Licensed under the MIT License.
 
 #include "init.h"
+#include <openenclave/bits/eeid.h>
 #include <openenclave/bits/sgx/sgxtypes.h>
+#include <openenclave/corelibc/string.h>
 #include <openenclave/enclave.h>
+#include <openenclave/internal/allocator.h>
 #include <openenclave/internal/calls.h>
+#include <openenclave/internal/eeid.h>
 #include <openenclave/internal/fault.h>
 #include <openenclave/internal/globals.h>
 #include <openenclave/internal/jump.h>
+#include <openenclave/internal/raise.h>
 #include <openenclave/internal/thread.h>
 #include "asmdefs.h"
 #include "td.h"
@@ -35,6 +40,37 @@ static void _check_memory_boundaries(void)
         oe_abort();
 }
 
+#ifdef OE_WITH_EXPERIMENTAL_EEID
+extern volatile const oe_sgx_enclave_properties_t oe_enclave_properties_sgx;
+extern oe_eeid_t* oe_eeid;
+
+static int _is_eeid_base_image(
+    const volatile oe_sgx_enclave_properties_t* properties)
+{
+    return properties->header.size_settings.num_heap_pages == 0 &&
+           properties->header.size_settings.num_stack_pages == 0 &&
+           properties->header.size_settings.num_tcs == 1;
+}
+
+static oe_result_t _eeid_patch_memory()
+{
+    oe_result_t r = OE_OK;
+
+    if (_is_eeid_base_image(&oe_enclave_properties_sgx))
+    {
+        uint8_t* enclave_base = (uint8_t*)__oe_get_enclave_base();
+        uint8_t* heap_base = (uint8_t*)__oe_get_heap_base();
+        oe_eeid_marker_t* marker = (oe_eeid_marker_t*)heap_base;
+        oe_eeid = (oe_eeid_t*)(enclave_base + marker->offset);
+
+        // Wipe the marker page
+        memset(heap_base, 0, OE_PAGE_SIZE);
+    }
+
+    return r;
+}
+#endif
+
 static void _initialize_enclave_image()
 {
     /* Relocate symbols */
@@ -42,6 +78,13 @@ static void _initialize_enclave_image()
     {
         oe_abort();
     }
+
+#ifdef OE_WITH_EXPERIMENTAL_EEID
+    if (_eeid_patch_memory() != OE_OK)
+    {
+        oe_abort();
+    }
+#endif
 
     /* Check that memory boundaries are within enclave */
     _check_memory_boundaries();
